wip 90% style for auth
